Add files via upload
Title screen is a space holder for a more finished one with animations.  The cursor for the main menu will show what the player is clicking on, and will go through the animation when something is clicked.  The character test had its spacing organized and should work in the system now.